Avoid compiler warning in TimeVal::as_double()
* glib/glibmm/timeval.h (TimeVal::as_double): Cast long int operands
to double explicitly, to avoid a conversion warning due to possible
loss of precision.  Reported by Alexander Shaduri, GNOME bug #617463.

+++ glib/glibmm/timeval.h
@@ -157,7 +157,7 @@ TimeVal& TimeVal::operator-=(long seconds)
 inline
 double TimeVal::as_double() const
 {
-  return tv_sec + ((double) tv_usec / (double) G_USEC_PER_SEC);
+  return double(tv_sec) + double(tv_usec) / double(G_USEC_PER_SEC);
 }
 
 inline
